deleting group modification

+++ src/main/java/com/backend/senior_backend/service/GroupsService.java
@@ -11,11 +11,16 @@ import org.springframework.validation.BindingResult;
 import com.backend.senior_backend.dto.GroupNameWithBudget;
 import com.backend.senior_backend.dto.GroupWithRoleDTO;
 import com.backend.senior_backend.dto.budgetAndExpensesDTO;
+import com.backend.senior_backend.models.Categories;
 import com.backend.senior_backend.models.Expenses;
 import com.backend.senior_backend.models.Groups;
+import com.backend.senior_backend.models.Participants;
+import com.backend.senior_backend.repositories.CategoriesRepository;
 import com.backend.senior_backend.repositories.ExpenseRepository;
 import com.backend.senior_backend.repositories.GroupsRepository;
 import com.backend.senior_backend.repositories.ParticipantsRepository;
+
+import jakarta.transaction.Transactional;
 
 @Service
 public class GroupsService {
@@ -34,6 +39,9 @@ public class GroupsService {
 
     @Autowired
     private ExpenseRepository expensesRepository;
+
+    @Autowired
+    private CategoriesRepository categoriesRepository;
     
     public Groups addGroup(Groups group, String phone, BindingResult bindingResult) {
         if (bindingResult.hasErrors()) {
@@ -94,12 +102,33 @@ public class GroupsService {
         return budgetAndExpenses;
     }
 
+
     public String deleteGroup(Long groupId) {
-        groupsRepository.deleteById(groupId);
+        // Ensure the group exists
+        Groups group = groupsRepository.findById(groupId).orElse(null);
+        if (group == null) {
+            return "Group not found!";
+        }
+
+        // Optional: Delete associated expenses
         List<Expenses> expenses = expensesRepository.findAllByCategoryIdGroupId(groupId);
-        expensesRepository.deleteAll(expenses);
+        if (!expenses.isEmpty()) {
+            expensesRepository.deleteAll(expenses);
+        }
+        List<Participants> participants = participantsRepository.findAllByGroupId(groupId.intValue());
+        if (!participants.isEmpty()){
+            participantsRepository.deleteAll(participants);
+        }
+        List<Categories> categories = categoriesRepository.findAllByGroupId(groupId);
+        if(!categories.isEmpty()){
+            categoriesRepository.deleteAll(categories);
+        }
+        groupsRepository.delete(group);
+
         return "Group deleted successfully";
     }
+
+    
 
     public String updateGroup(Long groupId, GroupNameWithBudget updatedGroup) {
         Groups group = groupsRepository.findById(groupId).orElse(null);
